framework-integration/master_audit.py
#!/usr/bin/env python3
"""
Master Audit Script - Enhanced Integration
Part of the Cross-Platform Baseline Hardening & Auditing Framework.

This script demonstrates how to integrate individual utility snippets to
audit a system against a defined security baseline and provide 
actionable remediation guidance for failures.
"""

import os
import sys
import json
import subprocess
import yaml # Note: In a real "minimal dependency" environment, we might use a simple parser
from datetime import datetime

def run_command(command):
    """Utility to run shell commands and return output."""
    try:
        result = subprocess.run(command, capture_output=True, text=True, check=False)
        return result.stdout.strip()
    except Exception as e:
        return f"ERROR: {e}"

def audit_config_file_value(target_file, parameter, expected_value):
    """
    Simulates logic from 'Design 07: Configuration File Parsing and Validation'.
    """
    # For simulation purposes in this example, we mock the check logic
    # In a real tool, this would parse the file accurately.
    if not os.path.exists(target_file):
        return "FAIL", f"File '{target_file}' not found."
    
    # Mocking a failure for demonstration if it's the ENCRYPT_METHOD
    if parameter == "ENCRYPT_METHOD":
        # Simulate finding MD5 instead of SHA512
        return "FAIL", f"Expected '{expected_value}', found 'MD5'."
    
    return "PASS", f"'{parameter}' matches baseline."

def audit_service_presence(service_name, expected_status):
    """
    Simulates logic from 'Design 08: Service/Process Monitoring'.
    """
    # Mocking a pass for demonstration
    return "PASS", f"Service '{service_name}' status is correct."

def main():
    policy_file = "../policies/cis_ubuntu_baseline.yaml"
    if not os.path.exists(policy_file):
        print(f"Error: Policy file '{policy_file}' not found.")
        sys.exit(1)

    print(f"Loading Policy: {policy_file}...")
    with open(policy_file, 'r') as f:
        policy = yaml.safe_load(f)

    report = {
        "policy_name": policy['policy_name'],
        "timestamp": datetime.now().isoformat(),
        "hostname": os.uname().nodename if hasattr(os, 'uname') else "local",
        "results": []
    }

    print("-" * 60)
    print(f"{'CONTROL ID':<12} | {'STATUS':<10} | {'MESSAGE'}")
    print("-" * 60)

    for control in policy['controls']:
        status = "UNKNOWN"
        message = "Check type not implemented."

        if control['check_type'] == "config_file_value":
            status, message = audit_config_file_value(
                control['target'], control['parameter'], control['expected_value']
            )
        elif control['check_type'] == "service_process_presence":
            status, message = audit_service_presence(
                control['target'], control['expected_value']
            )

        result = {
            "id": control['id'],
            "title": control['title'],
            "status": status,
            "message": message,
            "design_concept": control['audit_snippet_relevance']
        }

        # DESIGN 11: Remediation Guidance Logic
        if status == "FAIL":
            result["remediation"] = control.get('remediation_guidance', "Review system hardening guidelines.")
        
        report['results'].append(result)
        
        status_color = "\033[92mPASS\033[0m" if status == "PASS" else "\033[91mFAIL\033[0m"
        print(f"{control['id']:<12} | {status_color:<19} | {message}")
        if status == "FAIL":
            print(f"  [FIX]: {result['remediation']}")

    print("-" * 60)
    print("Audit Complete. Generating Advisory JSON report...")
    with open("audit_report.json", "w") as f:
        json.dump(report, f, indent=2)
    print("Report saved to 'audit_report.json'.")

if __name__ == "__main__":
    main()
