requirements.txt
pyyaml>=6.0
watchdog>=3.0.0
psutil>=5.9.0
